conformance/functionalityTesting/createKernelsInProgram.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>createKernelsInProgram conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

var globalWorkSize = new Int32Array(1);
var localWorkSize = new Int32Array(1);

var getKernels = function (webCLContext, kernelSource, webCLDevices, kernelName, kernelArgCount) {
    var webCLProgram = wtu.createProgram(webCLContext, kernelSource);
    wtu.build(webCLProgram, webCLDevices);
    try {
        webCLKernel = webCLProgram.createKernelsInProgram();
    } catch (e) {
        console.log("createKernelsInProgram threw exception: " + e.message);
        return null;
    }

    // FIXME This is adding a count to the nuber of test cases.
    shouldBeArrayOfType("webCLKernel", "WebCLKernel");

    for (i = 0; i < webCLKernel.length; i++) {
        if (!(webCLKernel[i].getInfo(webcl.KERNEL_CONTEXT) === webCLContext)) {
            console.log("webCLKernel[i].getInfo(webcl.KERNEL_CONTEXT) === webCLContext is false.");
            return null;
        }
        if (!(webCLKernel[i].getInfo(webcl.KERNEL_PROGRAM) === webCLProgram)) {
            console.log("webCLKernel[i].getInfo(webcl.KERNEL_PROGRAM) === webCLProgram is false.");
            return null;
        }
        if (!(webCLKernel[i].getInfo(webcl.KERNEL_NUM_ARGS) === stringify(kernelArgCount[i]))) {
            console.log("webCLKernel[i].getInfo(webcl.KERNEL_NUM_ARGS) === stringify(kernelArgCount[i]) is false");
            return null;
        }
        if (!(webCLKernel[i].getInfo(webcl.KERNEL_FUNCTION_NAME) === stringify(kernelName[i]))) {
            console.log("webCLKernel[i].getInfo(webcl.KERNEL_FUNCTION_NAME) === stringify(kernelName[i]) is false");
            return null;
        }
    }
    return {
        program: webCLProgram,
        kernelArray: webCLKernel
    };
}

try {
    debug("createKernelsInProgram functionality test case.");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLPlatforms = wtu.getPlatforms();
    var webCLDevices = wtu.getDevices(webCLPlatforms);
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var kernelSource, webCLKernel;
    var correct, correct_1, correct_2, input, output, resultObject;
    var count = SIZE;
    var data = new Float32Array(SIZE);
    var data_1 = new Float32Array(SIZE);
    var data_2 = new Float32Array(SIZE);
    var data_int = new Float32Array(SIZE);
    var data_float = new Float32Array(SIZE);
    var data_square = new Float32Array(SIZE);
    var result = new Float32Array(SIZE);
    var result_2 = new Float32Array(SIZE);

    /*
    *   Testing for "kernel : sample_copy_kernel"
    *   Copies int values from i/p to o/p.
    */
    for (i = 0; i < count; i++) {
        data[i] = Math.floor(Math.random() * 10);
    }

    kernelSource = wtu.readKernel("../../resources/kernels/sample_copy_kernel.cl");
    resultObject = getKernels(webCLContext, kernelSource, webCLDevices, ["sample_copy_kernel"], [2]);
    if (resultObject && resultObject.kernelArray) {
        var kernelCopy = resultObject.kernelArray[0];

        input = webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);
        output = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);
        webCLCommandQueue.enqueueWriteBuffer(input, true, 0, DATA_SIZE, data);

        kernelCopy.setArg(0, input);
        kernelCopy.setArg(1, output);

        globalWorkSize[0] = SIZE;
        localWorkSize[0] = kernelCopy.getWorkGroupInfo(webCLDevices[0], webcl.KERNEL_WORK_GROUP_SIZE);
        shouldBe("wtu.enqueueNDRangeKernel(webCLCommandQueue, kernelCopy, null , globalWorkSize, localWorkSize);", "true", true);
        webCLCommandQueue.enqueueReadBuffer(output, true, 0, DATA_SIZE, result);

        correct = 0;
        for (i = 0; i < count; i++) {
            if (result[i] === (data[i]))
                correct++;
        }
        console.log("Computed " + correct + "/" + count + " correct values for kernel: \"sample_copy_kernel\".");
        if (correct == count)
            testPassed("Test passed for all " + count + " values for kernel: \"sample_copy_kernel\".");
        else
            testFailed("Test failed. Computed " + correct + " / " + count + " correct values for kernel: \"sample_copy_kernel\".");
    } else
        testFailed("Failure in creating valid kernel objects for kernel: \"sample_copy_kernel\".");

    /*
    *   Testing for "kernel : square"
    *   Squares float value.
    */
    for (i = 0; i < count; i++) {
        data[i] = Math.random();
    }

    kernelSource = wtu.readKernel("../../resources/kernels/square.cl");
    resultObject = getKernels(webCLContext, kernelSource, webCLDevices, ["square"], [3]);
    if (resultObject && resultObject.kernelArray) {
        var kernelSquare = resultObject.kernelArray[0];

        input = webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);
        output = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);
        webCLCommandQueue.enqueueWriteBuffer(input, true, 0, DATA_SIZE, data);

        kernelSquare.setArg(0, input);
        kernelSquare.setArg(1, output);
        kernelSquare.setArg(2, count, WebCLKernelArgumentTypes.LONG);

        globalWorkSize[0] = SIZE;
        localWorkSize[0] = kernelSquare.getWorkGroupInfo(webCLDevices[0], webcl.KERNEL_WORK_GROUP_SIZE);
        shouldBe("wtu.enqueueNDRangeKernel(webCLCommandQueue, kernelSquare, null , globalWorkSize, localWorkSize);", "true", true);
        webCLCommandQueue.enqueueReadBuffer(output, true, 0, DATA_SIZE, result);

        correct = 0;
        for (i = 0; i < count; i++) {
            if (result[i].toPrecision(2) === (data[i] * data[i]).toPrecision(2))
                correct++;
        }
        console.log("Computed " + correct + "/" + count + " correct values for kernel \"square\".");
        if (correct == count)
            testPassed("Test passed for all " + count + " values for kernel: \"square\".");
        else
            testFailed("Test failed. Computed " + correct + " / " + count + " correct values for kernel: \"square\".");
    } else
        testFailed("Failure in creating valid kernel objects for kernel: \"square\".");

    /*
    *   Testing for "kernel : sample_single_kernel"
    *   Rounds off float to an int.
    */
    for (i = 0; i < count; i++) {
        data[i] = Math.random() * 10;
    }
    result = new Int32Array(SIZE);
    kernelSource = wtu.readKernel("../../resources/kernels/sample_single_kernel.cl");
    resultObject = getKernels(webCLContext, kernelSource, webCLDevices, ["sample_single_kernel"], [2]);
    if (resultObject && resultObject.kernelArray) {
        var kernelFloatToInt = resultObject.kernelArray[0];

        input = webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);
        output = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);
        webCLCommandQueue.enqueueWriteBuffer(input, true, 0, DATA_SIZE, data);

        kernelFloatToInt.setArg(0, input);
        kernelFloatToInt.setArg(1, output);

        globalWorkSize[0] = SIZE;
        localWorkSize[0] = kernelFloatToInt.getWorkGroupInfo(webCLDevices[0], webcl.KERNEL_WORK_GROUP_SIZE);
        shouldBe("wtu.enqueueNDRangeKernel(webCLCommandQueue, kernelFloatToInt, null , globalWorkSize, localWorkSize);", "true", true);
        webCLCommandQueue.enqueueReadBuffer(output, true, 0, DATA_SIZE, result);

        correct = 0;
        for (i = 0; i < count; i++) {
            if (result[i] == Math.floor(data[i]))
                correct++;
        }
        console.log("Computed " + correct + "/" + count + " correct values for kernel \"sample_single_kernel\".");
        if (correct == count)
            testPassed("Test passed for all " + count + " values for kernel: \"sample_single_kernel\".");
        else
            testFailed("Test failed. Computed " + correct + " / " + count + " correct values for kernel: \"sample_single_kernel\".");
    } else
        testFailed("Failure in creating valid kernel objects for kernel: \"sample_single_kernel\".");

    /*
    *   Testing for "two kernels in one"
    *   Consists of two kernels: to square and to roundoff a float to an int.
    */
    for (i = 0; i < count; i++) {
        data_square[i] = Math.random();
    }

    for (i = 0; i < count; i++) {
        data_float[i] = Math.random() * 10;
    }
    kernelSource = wtu.readKernel("../../resources/kernels/square.cl");
    kernelSource = kernelSource.concat(wtu.readKernel("../../resources/kernels/sample_single_kernel.cl"));
    resultObject = getKernels(webCLContext, kernelSource, webCLDevices, ["square" , "sample_single_kernel"], [3, 2]);
    shouldBe("webCLKernel.length", stringify(2), true);

    if (resultObject && resultObject.kernelArray) {

        var kernelPassed = 0;
        kernelSquare= resultObject.kernelArray[0];
        kernelFloatToInt = resultObject.kernelArray[1];

        var input_square = webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);
        var input_float = webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);
        var output_square = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);
        var output_int = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);
        webCLCommandQueue.enqueueWriteBuffer(input_square, true, 0, DATA_SIZE, data_square);
        webCLCommandQueue.enqueueWriteBuffer(input_float, true, 0, DATA_SIZE, data_float);

        kernelSquare.setArg(0, input_square);
        kernelSquare.setArg(1, output_square);
        kernelSquare.setArg(2, count, WebCLKernelArgumentTypes.LONG);

        kernelFloatToInt.setArg(0, input_float);
        kernelFloatToInt.setArg(1, output_int);

        globalWorkSize[0] = SIZE;
        localWorkSize[0] = kernelFloatToInt.getWorkGroupInfo(webCLDevices[0], webcl.KERNEL_WORK_GROUP_SIZE);
        shouldBe("wtu.enqueueNDRangeKernel(webCLCommandQueue, kernelSquare, null , globalWorkSize, localWorkSize);", "true", true);
        shouldBe("wtu.enqueueNDRangeKernel(webCLCommandQueue, kernelFloatToInt, null , globalWorkSize, localWorkSize);", "true", true);
        webCLCommandQueue.enqueueReadBuffer(output_square, true, 0, DATA_SIZE, result);
        webCLCommandQueue.enqueueReadBuffer(output_int, true, 0, DATA_SIZE, result_2);

        correct_1 = 0;
        for (i = 0; i < count; i++) {
            if (result[i].toPrecision(2) === (data_square[i] * data_square[i]).toPrecision(2))
                correct_1++;
        }
        if (correct_1 == count)
           kernelPassed++;

        correct_2 = 0;
        for (i = 0; i < count; i++) {
            if (result_2[i] == Math.floor(data_float[i]))
                correct_2++;
        }
        if (correct_2 == count)
           kernelPassed++;

        if (kernelPassed == 2)
            testPassed("Test passed for all " + count + " values for \"two kernels in one\" for kernels (\"square\" and \"sample_single_kernel\").");
        else
            testFailed("Test failed for \"two kernels in one\". Computed " + correct_1 + " / " + count + " correct values for kernel: \"square\" and " + correct_2 + " / " + count + " correct values for kernel: \"sample_single_kernel\".");

    } else
        testFailed("Test failed for \"two kernels in one\" for kernels (\"square\" and \"sample_single_kernel\") due to failure in validating created kernels.");

    /*
    *   Testing for "two kernels : sample_test, sample_test2"
    *   Converting float to int and to converting int to a float.
    */
    for (i = 0; i < count; i++) {
        data_int[i] = Math.floor(Math.random() * 10);
    }

    for (i = 0; i < count; i++) {
        data_float[i] = Math.random() * 10;
    }

    kernelSource = wtu.readKernel("../../resources/kernels/sample_two_kernels_in_one.cl");
    resultObject = getKernels(webCLContext, kernelSource, webCLDevices, ["sample_test", "sample_test2"], [2, 2]);
    shouldBe("webCLKernel.length", stringify(2), true);

    if (resultObject && resultObject.kernelArray) {

        var kernelPassed = 0;
        kernel_1= resultObject.kernelArray[0];
        kernel_2= resultObject.kernelArray[1];

        var input_1 = webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);
        var input_2 = webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);
        var output_1 = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);
        var output_2 = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);
        webCLCommandQueue.enqueueWriteBuffer(input_1, true, 0, DATA_SIZE, data_float);
        webCLCommandQueue.enqueueWriteBuffer(input_2, true, 0, DATA_SIZE, data_int);

        kernel_1.setArg(0, input_1);
        kernel_1.setArg(1, output_1);

        kernel_2.setArg(0, input_2);
        kernel_2.setArg(1, output_2);

        globalWorkSize[0] = SIZE;
        localWorkSize[0] = kernelFloatToInt.getWorkGroupInfo(webCLDevices[0], webcl.KERNEL_WORK_GROUP_SIZE);
        shouldBe("wtu.enqueueNDRangeKernel(webCLCommandQueue, kernel_1, null , globalWorkSize, localWorkSize);", "true", true);
        shouldBe("wtu.enqueueNDRangeKernel(webCLCommandQueue, kernel_2, null , globalWorkSize, localWorkSize);", "true", true);
        webCLCommandQueue.enqueueReadBuffer(output_1, true, 0, DATA_SIZE, result);
        webCLCommandQueue.enqueueReadBuffer(output_2, true, 0, DATA_SIZE, result_2);

        correct_1 = 0;
        for (i = 0; i < count; i++) {
            if (result[i] == data_1[i])
                correct_1++;
        }
        if (correct_1 == count)
            kernelPassed++;

        correct_2 = 0;
        for (i = 0; i < count; i++) {
            if (result_2[i] == (data_2[i]))
                correct_2++;
        }
        if (correct_2 == count)
            kernelPassed++;

        if (kernelPassed == 2)
            testPassed("Test passed for all " + count + " values for \"sample two kernels in one\" for kernels (\"sample_test\" and \"sample_test2\").");
        else
            testFailed("Test failed for \"sample two kernels in one\". Computed " + correct_1 + " / " + count + " correct values for kernel: \"sample_test\" and " + correct_2 + " / " + count + " correct values for kernel: \"sample_test2\".");

    } else
        testFailed("Test failed for \"sample two kernels in one\" for kernels (\"sample_test\" and \"sample_test2\") due to failure in validating created kernels.");

} catch(e) {
    testFailed("createKernelsInProgram throws an exception : " + e.message);
}
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
